Create lyric editor in composer.
