+++ osu.Game.Rulesets.Karaoke.Tests/Edit/TestSceneEditor.cs
@@ -2,7 +2,6 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using NUnit.Framework;
-using osu.Framework.Allocation;
 using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
 using osu.Game.Tests.Visual;
@@ -15,12 +14,5 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit
         protected override IBeatmap CreateBeatmap(RulesetInfo ruleset) => new TestKaraokeBeatmap(ruleset);
 
         protected override Ruleset CreateEditorRuleset() => new KaraokeRuleset();
-
-        [BackgroundDependencyLoader]
-        private void load()
-        {
-            // Catch clock because it will be used in lyric editor.
-            Dependencies.CacheAs(Clock);
-        }
     }
 }
